Move types into hostapi.v2.d.ts
